+++ src/app/returns-&-exchanges/page.tsx
@@ -2,7 +2,7 @@ import Footer from "@/src/components/Footer/Footer";
 import Header from "@/src/components/Header/Header";
 import ReturnsAndExchanges from "@/src/components/ReturnsAndExchanges/ReturnsAndExchanges";
 
-export default function page() {
+export default function ReturnAndExchange() {
   return (
     <div className="min-h-screen flex flex-col justify-between ">
       <Header />
